Refactor: ui 컴포넌트 앞으로 꺼내기
